spark/spark_app/requirements.txt
pyspark==3.4.0
kafka-python
elasticsearch
wordcloud
matplotlib
